Replace most custom Markdown plugins with native Starlight syntax
Edit the Markdown content to plain Starlight/HTML instead of preprocessing
it, cutting the custom plugin pipeline roughly in half:

- Admonitions now use Starlight's built-in aside variants and styling
  (info -> note, success/example -> tip, warning -> caution,
  failure/error -> danger, original titles preserved), dropping the
  Material-for-MkDocs restyle and extra variants.
- Collapsible ??? admonitions become plain <details>/<summary> HTML
  (which also renders on GitHub), with a light touch of CSS.
- {.screenshot} image markers are gone; content images get the same
  centred/capped styling via one CSS rule.
- Relative .md links are rewritten to page-relative URLs in the source
  (still checked by starlight-links-validator).
- The satteri-admonitions, satteri-image-attrs and satteri-md-links
  plugins are deleted along with their shared helpers and CSS.

Four small plugins remain, each with a concrete reason: code-snippet
includes (single source shared with rich-codex, which executes the
snippet files), content tabs (native <Tabs> needs MDX, which forbids the
HTML comments rich-codex reads), GitHub blockquote alerts (CHANGELOG and
CONTRIBUTING must keep GitHub-flavoured syntax to render on github.com),
and mermaid (no Starlight/Sätteri-compatible alternative).
