+++ CONTRIBUTING.md
@@ -68,7 +68,12 @@ npm run dev
 This starts a local dev server with hot reload. `npm run build` builds the full
 static site (including link validation) into `docs/dist/`.
 
-Most pages live in `docs/src/content/docs/`. The landing page, changelog and
+Most pages live in `docs/src/content/docs/` and use standard
+[Starlight Markdown](https://starlight.astro.build/guides/authoring-content/)
+(asides, plain `<details>` blocks, etc.), plus a couple of small local
+extensions: `:::tab[Label]` directives for content tabs and
+`{% include "..." %}` inside code fences to pull in runnable snippets from
+`docs/code_snippets/` (see `docs/plugins/`). The landing page, changelog and
 contributing pages are generated from the repo-root `README.md`, `CHANGELOG.md`
 and `CONTRIBUTING.md` files by `docs/scripts/prepare-content.mjs`. The terminal
 screenshots are auto-generated by [rich-codex](https://github.com/ewels/rich-codex)
