--- docs/plugins/satteri-image-attrs.mjs
@@ -1,41 +1,0 @@
-// Minimal port of the Python-Markdown `attr_list` extension for the syntax the
-// docs actually use: a `{.class}` / `{ .class }` marker immediately following
-// an image or link adds those classes to the element, e.g.:
-//
-//     ![](../images/example.svg){.screenshot}
-//     [Watch on YouTube](https://youtu.be/...){ .md-button }
-const ATTR_RE = /^\{\s*((?:\.[\w-]+\s*)+)\}/;
-
-export function satteriImageAttrs() {
-  return {
-    name: 'rich-click-image-attrs',
-    paragraph(node, ctx) {
-      // Paragraph-shaped nodes generated by other plugins may have no children.
-      if (!node.children) return;
-      for (let i = 0; i < node.children.length - 1; i++) {
-        const target = node.children[i];
-        if (target.type !== 'image' && target.type !== 'link') continue;
-        const marker = node.children[i + 1];
-        if (marker.type !== 'text') continue;
-        const match = ATTR_RE.exec(marker.value);
-        if (!match) continue;
-
-        const classes = match[1]
-          .split(/\s+/)
-          .filter(Boolean)
-          .map((cls) => cls.replace(/^\./, ''))
-          .join(' ');
-        const data = { ...target.data };
-        data.hProperties = { ...data.hProperties, class: classes };
-        ctx.setProperty(target, 'data', data);
-
-        const rest = marker.value.slice(match[0].length);
-        if (rest.length > 0) {
-          ctx.setProperty(marker, 'value', rest);
-        } else {
-          ctx.removeNode(marker);
-        }
-      }
-    },
-  };
-}
