+++ docs/plugins/satteri-mermaid.mjs
@@ -2,7 +2,7 @@
 // that are rendered client-side by the script in src/components/Header.astro.
 // Running as an mdast plugin means the block never reaches the syntax
 // highlighter, mirroring how mkdocs-material handled mermaid superfences.
-import { element } from './admonition-common.mjs';
+import { element } from './mdast-helpers.mjs';
 
 export function satteriMermaid() {
   return {
